Int64 instead of Int

+++ src/types.jl
@@ -32,7 +32,7 @@ Tuple4(phone::Int32, hmm_state::Int32, pdf_class::Int32) = Tuple4(phone, hmm_sta
 
 
 struct TransitionModel{T<:AbstractFloat}
-	topo::Dict{Int, TopologyEntry}
+	topo::Dict{Int64, TopologyEntry}
 	tuples::Vector{Tuple4}
 	log_probs::Vector{T}
 end
